ui test file for python lsp server
