+++ ui-tests/python-lsp-server-test/python_lsp_server.spec.ts-snapshots/lsp-popup-fully-initialized-linux.png
